requirements.txt
absl-py==0.9.0
apturl==0.5.2
asn1crypto==0.24.0
blis==0.4.1
Brlapi==0.6.6
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.16
de-core-news-sm==2.2.5
decorator==4.4.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
easydict==1.9
en-core-web-sm==2.2.5
future==0.18.2
fvcore==0.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
httplib2==0.9.2
idna==2.9
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.5.0
json-tricks==3.15.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
llvmlite==0.31.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==3.2.1
murmurhash==1.0.2
netifaces==0.10.4
networkx==2.4
numba==0.48.0
numpy==1.18.2
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.2.0.32
pandas==1.0.3
panopticapi==0.1
pexpect==4.2.1
Pillow==7.1.1
plac==1.1.3
portalocker==1.5.2
preshed==3.0.2
progress==1.5
protobuf==3.11.3
psutil==5.4.2
pyarmor==5.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycocotools==2.0
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyrealsense2==2.32.1.1331
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-xapp==1.0.0
python-xlib==0.20
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
SecretStorage==2.3.1
selenium==3.141.0
sentencepiece==0.1.85
setproctitle==1.1.10
Shapely==1.6.4
simplejson==3.13.2
six==1.14.0
spacy==2.2.3
spark-parser==1.8.9
srsly==1.0.2
stanza==1.0.0
supervisor==4.1.0
system-service==0.3
systemd-python==234
tabulate==0.8.6
tensorboard==2.1.0
tensorboardX==1.6
termcolor==1.1.0
thinc==7.3.1
torch==1.4.0
torchtext==0.5.0
torchvision==0.5.0
tqdm==4.43.0
transforms==0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uncompyle6==3.6.5
urllib3==1.25.8
usb-creator==0.3.3
wadllib==1.3.2
wasabi==0.6.0
Werkzeug==1.0.0
xdis==4.2.4
xkit==0.0.0
yacs==0.1.6
zipp==3.1.0
zope.interface==4.3.2
